AIServiceHub_Mgmt/AIServiceHub/Patch_NetVLAD/requirements.txt
numpy 
pytorch-gpu 
torchvision 
natsort 
tqdm 
pillow 
scikit-learn 